Integrate the Screen-Space Effects and NPE Sprite2D demos

+++ native/include/bblite/pal.hpp
@@ -81,7 +81,7 @@ double monotonic_milliseconds();
 // Browser-facing `performance.now()`. In ordinary runs this is the same
 // monotonic clock; fixed-delta captures advance it deterministically.
 double performance_milliseconds();
-void advance_performance_milliseconds(float delta_ms);
+void advance_performance_milliseconds(double delta_ms);
 /**
  * The process's resident working set in bytes, or 0 where the platform
  * has no query. Read by the BBLITE_MEM_PROFILE frame line, which is how a
